refactor: move files to evaluators/
